Switch to C++ oeedger8r Benefits - Reduced barrier to contribution - Avoids tooling issues; just needs C++ compiler - Easier to add features due to use of a common programming language
Remove dependency on nodejs, esy
Benefits
- Significantly reduced SDK build times on fresh Windows machines.
  Previously took additional 20-30 mins to build Ocaml compiler and then oeedger8r
- No need for administrator prompt to build OE SDK
- Windows Pipeline reduces from 48 minutes to 24 minutes

Build oeedger8r as external project.

This ensures that ARM binary is not generated for OE_TrustZone
build of the SDK which requires the edger8r to be run on the
host x64 machine.

Make following tests robust to oeedger8r search path processing order
- tests/memory (memory.edl)
- tests/syscall/epoll (epoll.edl)

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>